requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
numpy>=1.24.0
pillow>=9.5.0

# API and web framework
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Machine Learning and AI
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0
safetensors>=0.3.1
peft>=0.4.0

# Data processing and utilities
pandas>=2.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pyyaml>=6.0

# Testing and development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0

# Optional: GPU support
# cuda-python>=12.0.0  # Uncomment if using NVIDIA GPU 